+++ packages/zowe-explorer/src/utils/ProfilesUtils.ts
@@ -49,7 +49,7 @@ export async function errorHandling(errorDetails: any, label?: string, moreInfo?
     // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
     globals.LOG.error(`${errorDetails}\n` + JSON.stringify({ errorDetails, label, moreInfo }));
 
-    if (errorDetails.mDetails !== undefined) {
+    if (errorDetails?.mDetails !== undefined) {
         httpErrCode = errorDetails.mDetails.errorCode;
         // open config file for missing hostname error
         const msg = errorDetails.toString();